Fix handling of different byte orders on SDL_Surface when converting to GL texture
SDL_ttf is generating argb32 images, while loaded images are rgba to start with.
With this change, we'll now convert the SDL surface to rgba32 if it's not that
format already.

Note that previously text was working with :bgra, which is :argb in big-endian
format. That's why I went with rgba32 vs rgba8888, which would have the wrong
endianness. I tested all of this out to come to that conclustion.

Tested this change and both text and images are loading with the proper colors
now.