Cut image surface on load
Adds 4 optional params (x y w h) when loading image resources, to put
only a part of the image into the texture.

This is how to tile a part of an image:

```lisp
(defsketch tile
    ((width 400) (height 400))
  (let ((res (load-resource "100x100.png" :x 0 :y 0 :w 50 :h 50)))
    (with-pen (make-pen :fill (crop res 0 0 100 100))
      (rect 0 0 400 400))))
```

Here, only the upper left 50x50 part of the image is loaded. As
`crop` (unintuivitely?) tiles the image when width or height are
larger than the ones of the image, the intended effect is achieved.