Simplify
In the lambda list the default value form won't be called if the argument was
passed. Technically the slot-value of the instance was used instead of the value
passed. If PREPARE is called from the INITIALIZE-INSTANCE - we just passed these
values to the initialize instance, so they are the same. If PREPARE is called from the
UPDATE-INSTANCE-FOR-REDEFINED-CLASS, initargs are empty, so the slot-value wouldn't be
called anyways.