Optimize COPY-BUFFER (also gets rid of warnings)
COPY-BUFFER is only used when drawing a figure.
SRC-OFFSET and DST-OFFSET are not used, so it is safe to assume they are fixnums (they are always 0).
LENGTH is the length of the static-vector which holds the cached buffer data.
By ANSI it is bounded by the ARRAY-DIMENSION-LIMIT which is guaranteed to be a fixnum, so LENGTH is a fixnum too.