Take the current PEN-WEIGHT into account when drawing points
Also, use CIRCLE instead of RECT -- that's what users will expect to see
with bigger values of PEN-WEIGHT anyways.

Note that this will change where points are rendered: circles are
centered around the specified XY, while rectangles use that XY as its
top-left corner (or bottom-left when :Y-AXIS is :UP).