Revise input-handling interface.
There's now a tree of 'on-mouse-' functions, the user can
hook into them at any point. 'on-mouse-button' propagates
to 'on-mouse-left' which propagates to 'on-mouse-left-up'
and 'on-mouse-left-down'.

Keeps 'on-click', since that's a simple and common case.

Also changes the 'state' parameter 'on-key' so that it can have
values ':up' or ':down', to be consistent with the mouse interface.